move vm console watching code to vm_console.go
